+++ client/src/config/config.js
@@ -1,4 +1,5 @@
 module.exports = {
 	//backendEndpoint: "http://192.168.1.7:4001"
-	backendEndpoint: "http://192.168.1.3:4001"
+	//backendEndpoint: "http://192.168.1.3:4001"
+	backendEndpoint: "http://192.168.10.115:4001"
 }