+++ client/src/utils/tools.js
@@ -1,0 +1,6 @@
+
+//TODO: move
+export function runTool(toolName) {
+	console.log("TODO: runTool " + toolName);
+}
+//